backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0

SQLAlchemy==2.0.13
# psycopg2==2.9.6 
psycopg2-binary==2.9.6

python-dotenv==1.0.0